Phase 2 of #827: extract Alpine factory from rule_form
Move the rule_form Alpine factory out of the rule_form_scripts.go raw
Go string into a dedicated static/js/admin/components/rule_form.js
module, following the convention codified in #828.

The factory now reads the existing rule (edit mode only) from a
@templ.JSONScript("rule-form-data", p.Rule) tag and reads the isEdit
flag from a data-is-edit attribute on the x-data root. The script tag
loads synchronously at the top of the templ component so the
Alpine.data() registration runs before alpine:init fires.
